+++ demo/site-pages/src/pages/api/status.tsx
@@ -1,0 +1,7 @@
+import { NextApiRequest, NextApiResponse } from "next";
+
+const Status = (req: NextApiRequest, res: NextApiResponse): void => {
+    res.statusCode = 200;
+    res.send("OK");
+};
+export default Status;