+++ install.sh
@@ -36,9 +36,16 @@ ln -sf ../api/schema.gql ./demo/site/schema.gql
 ln -sf ../api/block-meta.json ./demo/site/block-meta.json
 ln -sf ../../api/src/comet-config.json ./demo/site/src/comet-config.json
 
+# site-pages DEMO
+ln -sf ../../.env ./demo/site-pages/.env
+ln -sf ../api/schema.gql ./demo/site-pages/schema.gql
+ln -sf ../api/block-meta.json ./demo/site-pages/block-meta.json
+ln -sf ../../api/src/comet-config.json ./demo/site-pages/src/comet-config.json
+
 # Lang install
 sh ./demo/admin/intl-update.sh
 sh ./demo/site/intl-update.sh
+sh ./demo/site-pages/intl-update.sh
 
 # Build the packages CLI and eslint-plugin to be used for dev startup
 pnpm --filter '@comet/cli' --filter '@comet/eslint-plugin' run build