+++ packages/admin/admin-react-select/CHANGELOG.md
@@ -1,6 +1,5 @@
 # @comet/admin-react-select
 
-<<<<<<< HEAD
 ## 7.0.0-beta.3
 
 ### Patch Changes
@@ -173,7 +172,7 @@
 -   Updated dependencies [b0249e3bc]
 -   Updated dependencies [92eae2ba9]
     -   @comet/admin@7.0.0-beta.0
-=======
+
 ## 6.15.1
 
 ### Patch Changes
@@ -189,7 +188,6 @@
 -   Updated dependencies [0654f7bce]
     -   @comet/admin-icons@6.15.0
     -   @comet/admin@6.15.0
->>>>>>> main
 
 ## 6.14.1