+++ packages/admin/admin-react-select/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @comet/admin-react-select
 
+## 7.0.0-beta.3
+
+### Patch Changes
+
+-   Updated dependencies [ce5eaede2]
+    -   @comet/admin@7.0.0-beta.3
+    -   @comet/admin-icons@7.0.0-beta.3
+
 ## 7.0.0-beta.2
 
 ### Patch Changes
@@ -164,6 +172,22 @@
 -   Updated dependencies [b0249e3bc]
 -   Updated dependencies [92eae2ba9]
     -   @comet/admin@7.0.0-beta.0
+
+## 6.15.1
+
+### Patch Changes
+
+-   @comet/admin@6.15.1
+-   @comet/admin-icons@6.15.1
+
+## 6.15.0
+
+### Patch Changes
+
+-   Updated dependencies [406027806]
+-   Updated dependencies [0654f7bce]
+    -   @comet/admin-icons@6.15.0
+    -   @comet/admin@6.15.0
 
 ## 6.14.1